third_party/blink/web_tests/http/tests/origin_trials/webexposed/webcomponentsv0-basic-import-disabled.html

<!DOCTYPE html>
<meta charset="utf-8">

<link id="target" rel="import" href="resources/html-imports-hello.html">

<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<script>
// Behavior is different depending on if the runtime flag is enabled.
if (self.internals.runtimeFlags.htmlImportsEnabled) {
	// Runtime flag is enabled, so import will actually work. Explicitly call
	// done() to avoid timeout while test framework waits for a test to run.
	done();
} else {
  test(() => {
  	  assert_equals(target.import, undefined);
    }, 'Basic html file should not be imported');
}
</script>
